% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_top_prevalence.R
\name{calculate_top_prevalence}
\alias{calculate_top_prevalence}
\title{A tool to search for taxa's prevalence and mean relative abundance from your phyloseq object by the top n most prevalent taxa}
\usage{
calculate_top_prevalence(physeq, n = 20, group_var = NULL)
}
\arguments{
\item{physeq}{A phyloseq object.}

\item{n}{The number of top taxa to show (sorted by prevalence).}

\item{group_var}{The metadata column name to group by (e.g., "body.site").}
}
\value{
A data frame with columns for group, taxon, taxonomic ranks, mean relative abundance, and prevalence.
}
\description{
This function calculates the mean relative abundance and prevalence of each taxon
from a phyloseq object, grouped by a specified metadata column and sorted by prevalence.
It includes both the hash ID and associated taxonomic ranks.
}
\details{
Calculate Mean Relative Abundance and Prevalence by Group
}
