% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_taxonomy_info.R
\name{find_taxonomy_info}
\alias{find_taxonomy_info}
\title{Find specific hash id in a phyloseq object by searching the name of a taxa}
\usage{
find_taxonomy_info(physeq, tax_level, tax_name)
}
\arguments{
\item{physeq}{A phyloseq object containing the taxonomic table.}

\item{tax_level}{The taxonomic level to search (e.g., "Genus").}

\item{tax_name}{The name within the specified taxonomic level to find.}
}
\value{
A data frame with taxonomic information for the matching taxa.
}
\description{
Searches the taxonomic table of a phyloseq object for a specified taxonomic level and name.
}
\details{
Find Taxonomic Information
}
\examples{
\dontrun{
  # Assuming 'physeq' is a phyloseq object loaded in your environment
find_taxonomy_info(physeq, "Genus", "Acinetobacter")
}

}
\author{
Elyse Barker
}
