% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_taxa.R
\name{find_unique_taxa}
\alias{find_unique_taxa}
\title{A tool to search for taxa that is unique to a specific sample type and not found in any other sample type in the phyloseq table}
\usage{
find_unique_taxa(physeq_obj, sample_type_column, target_sample_type)
}
\arguments{
\item{physeq_obj}{A phyloseq object.}

\item{sample_type_column}{The column title that has the sample type information (e.g "body.site").}

\item{target_sample_type}{The sample type of interest (e.g., "Gill").}
}
\value{
A data frame with columns for taxon, taxonomic ranks, mean relative abundance, and prevalence that are all unique to the specific sample type.
}
\description{
This function calculates the mean relative abundance and prevalence of each unique taxon.
automatically calculates mean relative abundance within the function so input a phyloseq object with raw counts
It includes both the hash ID and associated taxonomic ranks.
}
\details{
Find unique taxa in a specific sample type that is not present in any other sample type
}
