% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_to_taxa.R
\name{hash_to_taxa}
\alias{hash_to_taxa}
\title{Get Taxonomic Information for a Given Hash ID}
\usage{
hash_to_taxa(physeq, hash_id)
}
\arguments{
\item{physeq}{A phyloseq object containing taxonomic information.}

\item{hash_id}{The hash ID of the taxon of interest.}
}
\value{
A data frame containing the taxonomic ranks for the specified hash ID.
}
\description{
This function retrieves taxonomic information for a specific hash ID from a given phyloseq object.
}
