% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_abundance_prevalence_microbiome.R
\name{MyMicrobiomeTools}
\alias{MyMicrobiomeTools}
\title{"A tool to search for a specific taxa's prevalence and mean relative abundance from your phyloseq object"}
\usage{
calculate_abundance_prevalence <- function(physeq, taxon_id, category)
}
\description{
calculates the mean relative abundance and prevalence from a phyloseq object
You input the hash id of the taxa of interest, your phyloseq object, and what category you want to look at
this function automatically calculates to relative abundances so input phyloseq object needs to be raw counts
need to have one column in the metadata be only one body site/environment of interest
}
\details{
calculate_abundance_prevalence
}
\examples{
# example code
#"body.site" column of the metadata only has the body site "gill". You can filter your phyloseq object beforehand to just have one site of interest.
result <- calculate_abundance_prevalence(gill_not_rel, "8d9fa267695027600ad8cf2eaca55c9c", "body.site")
}
\author{
Elyse Barker
}
